+++ Sources/VCore/Extensions/SwiftUI/Color/Color.PrimaryInverted.swift
@@ -1,0 +1,50 @@
+//
+//  Color.PrimaryInverted.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 24.01.24.
+//
+
+import SwiftUI
+
+// MARK: - Color Primary Inverted
+extension Color {
+    /// `Color` that can be uses for inverted primary content.
+    ///
+    ///     Text("Lorem ipsum")
+    ///         .foregroundStyle(Color.primaryInverted)
+    ///         .padding()
+    ///         .background(content: { Color.primary })
+    ///
+    public static let primaryInverted: Color = {
+#if os(iOS)
+        .dynamic(light: .white, dark: .black)
+#elseif os(macOS)
+        .dynamic(light: .white, dark: .black)
+#elseif os(tvOS)
+        .dynamic(light: .white, dark: .black)
+#elseif os(watchOS)
+        .black
+#elseif os(visionOS)
+        .black
+#endif
+    }()
+}
+
+// MARK: - Preview
+#if DEBUG
+
+#Preview(body: {
+    VStack(content: {
+        Text("Lorem ipsum")
+            .foregroundStyle(.primary)
+            .padding()
+
+        Text("Lorem ipsum")
+            .foregroundStyle(Color.primaryInverted)
+            .padding()
+            .background(content: { Color.primary })
+    })
+})
+
+#endif